Use new configuration format
According to https://atom.io/docs/api/v0.175.0/Config the recommended
format for configurations is to use js-schema.  That document also says
that "The settings view will use the title and description keys to
display your config setting in a readable way"

This makes it easier for users to understand what each configuration
does.